refactor: split out a test util module
In the Bloop code base the integration tests also actually use this
trait. It's only a single file so in reality we _could_ duplicate it in
Bloop, but I figured we could also simply publish it as a mini-module
that Bloop could just use in its integration tests. It's not really
cleaned up or anything special, so I'm unsure if anything other than
Bloop will actually use this.

Bloop would then simply add:

```scala
Dependencies.bloopConfigTestUtil % Test
```

To the plugin modules.