refactor: remove publishing for 2.11
Note that this doesn't mean you can't use Bloop with 2.11, it just means
that an application using 2.11 can't use the config library anymore to
generate the config. Seeing that everything I know of is probably on
2.13, I think we're pretty safe to drop 2.11. It'd be nice to also add
in Scala 3, but we use too old of a jsoniter version for this. We're
locked on an older version to support Java 8. However, I'm assuming
we'll change that soon as well, and then we can bump this and cross
publish to 3 as well.